+++ src/lib/hierarchy-operations/lineage-constraints.js
