+++ src/lib/hierarchy-operations/lineage-manipulation.js
